devices almost done

+++ windopen/rpyc_server/rpyc_service.py
@@ -13,8 +13,8 @@ class MTUService(rpyc.Service):
         The method runs when a new connection is accepted.
         Here is registered each new RTU
         """
-        log.info('Connecting device: %s', self._conn._config['connid'])
-        self.register_new_device(self._conn)
+        #log.info('Connecting device: %s', self._conn._config['connid'])
+        return self.register_new_device(self._conn)
 
     def on_disconnect(self):
         # cod care ruleaza dupa ce o conexiune este inchisa
@@ -23,8 +23,16 @@ class MTUService(rpyc.Service):
         for uuid, conn in self.conns.iteritems():
             log.info('DISCONECT uuid: %s  connid: %s', uuid, conn._config['connid'])
             if conn._config['connid'] == self._conn._config['connid']:
+                device = Device.objects.get(uuid = uuid)
+                if device:
+                    log.info('Change status to False to device %s: ', uuid)
+                    device.active = False
+                    device.last_seen = datetime.now()
+                    device.save()
                 unreg_device = UnregisteredDevice.objects.get(uuid=uuid)
-                unreg_device.delete()
+                if unreg_device:
+                    log.info('delete the unregistred device %s: ', uuid)
+                    unreg_device.delete()
                 try:
                     self.conns.pop(uuid)
                 except Exception as err:
@@ -50,11 +58,11 @@ class MTUService(rpyc.Service):
         return False
     
     def register_new_device(self, conn):
-        uuid = self._conn.root.get_uuid()
+        uuid = conn.root.get_uuid()
         log.info('new_uuid: %s', uuid)
         # search through the paired devices
-        dvs = Device.objects.filter(uuid=uuid)
-        log.debug('paired devices: %s', dvs)
+        dvs = Device.objects.get(uuid=uuid)
+        log.info('######### devices already: %s',dvs)
         if dvs:
             dvs.active = True
             dvs.last_seen = datetime.now()
@@ -63,4 +71,5 @@ class MTUService(rpyc.Service):
             new_device = UnregisteredDevice(uuid=uuid)
             new_device.save()
         self.conns[uuid] = self._conn
-        log.info('connections: %s', self.conns)
+        log.info('registered: %s', uuid)
+        return True
